requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
Authlib==1.3.1
certifi==2023.7.22
cffi==1.17.1
charset-normalizer==3.2.0
click==8.1.8
cryptography==44.0.0
dataclasses-json==0.6.7
distlib==0.3.8
Django==4.2.5
django-scheduler==0.10.1
djangorestframework==3.14.0
exceptiongroup==1.2.2
filelock==3.15.3
frozenlist==1.4.0
grpcio==1.70.0
grpcio-health-checking==1.70.0
grpcio-tools==1.70.0
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.1
icalendar==5.0.11
idna==3.4
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.76
langsmith==0.4.28
marshmallow==3.26.1
multidict==6.0.4
mypy_extensions==1.1.0
numpy==2.0.2
openai==0.28.0
orjson==3.11.3
packaging==25.0
pdf2image==1.16.3
pillow==11.3.0
pillow_heif==1.1.0
platformdirs==4.2.2
protobuf==5.29.3
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.10.1
pydantic_core==2.27.2
python-dateutil==2.8.2
python-dotenv==1.1.1
pytz==2023.3.post1
PyYAML==6.0.2
requests==2.32.5
requests-toolbelt==1.0.0
sentry-sdk==2.20.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlparse==0.4.4
tenacity==9.1.2
tqdm==4.66.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
urllib3==2.0.4
validators==0.34.0
virtualenv==20.26.2
weaviate-client==4.10.4
yarl==1.9.2
zstandard==0.25.0
